Fixed bugs on removing unneccessary libraries, removing mock test, and optimizing process meta
Added / for consistency in chatgpt-log-ts. Removed unneccessary comments
