+++ data/2024-10-31.jsonl
@@ -1,0 +1,9 @@
+{"requestId":"m2xg5yu9ar7x24dn4s","sessionId":"m2xg5yu9ar7x24dn4s","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gemini-1.5-flash","temperature":1,"maxOutputTokens":200},"response":{"text":"","finishReason":"STOP","completionTime":"2024-10-31T15:15:13.503Z","tokenCount":71,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","args":{"location":"Ikorodu, Lagos"},"result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{}}
+{"requestId":"m2xg60vdakaripl0x5p","sessionId":"m2xg5yu9ar7x24dn4s","request":{"prompt":"Hello, World!","model":"gemini-1.5-mini","temperature":0.7,"maxOutputTokens":150},"response":{"text":"It's sunny in Lagos!","finishReason":"stop","completionTime":"2024-10-31T15:15:13.513Z","tokenCount":120,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","args":"{\"location\":\"Lagos\"}","exitCode":0,"result":"Sunny, 22°C"}],"meta":{"totalTokenCount":120,"inputTokenCost1k":0.6,"outputTokenCost1k":0.8,"triggerSource":"API","outputMode":"streaming","userId":"hashed_user_id","country":"Nigeria","operatingSystem":"mac os","shell":"/bin/bash","userTimeZone":"WAT","memory":16,"machineId":"machine_id","env":"production"}}
+{"requestId":"m2xg5zu7afk1uj5if06","sessionId":"m2xg5zu7afk1uj5if06","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gpt-4o","temperature":0.7,"topP":0,"functionCalls":[{"name":"getWeather","description":"Gets the weather information for a location","parameters":{"type":"object","properties":{"location":{"type":"string","description":"The name of the city to get the weather for"}},"required":["location"]}}],"maxTokens":150,"errorReason":""},"response":{"finishReason":"function_call","completionTime":"2024-10-31T15:15:14.408Z","tokenCount":85,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{}}
+{"requestId":"m2xg82y5hpb0e2f9cfr","sessionId":"m2xg82y5hpb0e2f9cfr","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gemini-1.5-flash","temperature":1,"maxOutputTokens":200},"response":{"text":"","finishReason":"STOP","completionTime":"2024-10-31T15:16:51.108Z","tokenCount":71,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","args":{"location":"Ikorodu, Lagos"},"result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{}}
+{"requestId":"m2xg846k6zrfk5vo6y8","sessionId":"m2xg82y5hpb0e2f9cfr","request":{"prompt":"Hello, World!","model":"gemini-1.5-mini","temperature":0.7,"maxOutputTokens":150},"response":{"text":"It's sunny in Lagos!","finishReason":"stop","completionTime":"2024-10-31T15:16:51.116Z","tokenCount":120,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","args":"{\"location\":\"Lagos\"}","exitCode":0,"result":"Sunny, 22°C"}],"meta":{"totalTokenCount":120,"inputTokenCost1k":0.6,"outputTokenCost1k":0.8,"triggerSource":"API","outputMode":"streaming","userId":"hashed_user_id","country":"Nigeria","operatingSystem":"mac os","shell":"/bin/bash","userTimeZone":"WAT","memory":16,"machineId":"machine_id","env":"production"}}
+{"requestId":"m2xg837vvjggdbz8w4h","sessionId":"m2xg837vvjggdbz8w4h","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gpt-4o","temperature":0.7,"topP":0,"functionCalls":[{"name":"getWeather","description":"Gets the weather information for a location","parameters":{"type":"object","properties":{"location":{"type":"string","description":"The name of the city to get the weather for"}},"required":["location"]}}],"maxTokens":150,"errorReason":""},"response":{"finishReason":"function_call","completionTime":"2024-10-31T15:16:51.458Z","tokenCount":85,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{}}
+{"requestId":"m2xgxlitbnvc81ymx98","sessionId":"m2xgxlitbnvc81ymx98","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gpt-4o","temperature":0.7,"topP":0,"functionCalls":[{"name":"getWeather","description":"Gets the weather information for a location","parameters":{"type":"object","properties":{"location":{"type":"string","description":"The name of the city to get the weather for"}},"required":["location"]}}],"maxTokens":150,"errorReason":""},"response":{"finishReason":"function_call","completionTime":"2024-10-31T15:36:42.861Z","tokenCount":85,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{}}
+{"requestId":"m2xgxlc6x5insno121b","sessionId":"m2xgxlc6x5insno121b","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gemini-1.5-flash","temperature":1,"maxOutputTokens":200},"response":{"text":"","finishReason":"STOP","completionTime":"2024-10-31T15:36:42.888Z","tokenCount":71,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","args":{"location":"Ikorodu, Lagos"},"result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{}}
+{"requestId":"m2xgxnrle2g9x5fo9do","sessionId":"m2xgxlc6x5insno121b","request":{"prompt":"Hello, World!","model":"gemini-1.5-mini","temperature":0.7,"maxOutputTokens":150},"response":{"text":"It's sunny in Lagos!","finishReason":"stop","completionTime":"2024-10-31T15:36:42.897Z","tokenCount":120,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","args":"{\"location\":\"Lagos\"}","exitCode":0,"result":"Sunny, 22°C"}],"meta":{"totalTokenCount":120,"inputTokenCost1k":0.6,"outputTokenCost1k":0.8,"triggerSource":"API","outputMode":"streaming","userId":"hashed_user_id","country":"Nigeria","operatingSystem":"mac os","shell":"/bin/bash","userTimeZone":"WAT","memory":16,"machineId":"machine_id","env":"production"}}
