+++ data/2024-11-01.jsonl
@@ -1,0 +1,2 @@
+{"requestId":"m2yo09jdlkf0pn819r","sessionId":"m2yo09jdlkf0pn819r","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gpt-4o","temperature":0.7,"topP":0,"functionCalls":[{"name":"getWeather","description":"Gets the weather information for a location","parameters":{"type":"object","properties":{"location":{"type":"string","description":"The name of the city to get the weather for"}},"required":["location"]}}],"maxTokens":150,"errorReason":""},"response":{"finishReason":"function_call","completionTime":"2024-11-01T11:42:31.094Z","tokenCount":85,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{"totalTokenCount":85,"inputTokenCost1k":0.025,"outputTokenCost1k":0.1,"triggerSource":"","userId":"7f6bfa5b-8702-4602-ab24-a1dd454e061a","country":"","operatingSystem":"Darwin 21.6.0","shell":"/bin/zsh","userTimeZone":"Africa/Lagos","memory":0,"machineId":"bdb7e427269f249517f77a7db88829191951f3b5a19e51e4b24e806a2a6501b8","env":"test"}}
+{"requestId":"m2yo09bk2w8jqwtb1sa","sessionId":"m2yo09bk2w8jqwtb1sa","request":{"prompt":"What's the weather in Ikorodu, Lagos?","model":"gemini-1.5-flash","temperature":1,"maxOutputTokens":200},"response":{"text":"","finishReason":"STOP","completionTime":"2024-11-01T11:42:31.742Z","tokenCount":71,"status":200,"errorReason":""},"functionCalls":[{"name":"getWeather","args":{"location":"Ikorodu, Lagos"},"result":{"location":"Ikorodu, Lagos","temperature":"22°C","condition":"Sunny"}}],"meta":{"totalTokenCount":71,"triggerSource":"","userId":"aeee4c80-a984-49b8-ad34-5ea431ddd2b4","country":"","operatingSystem":"Darwin 21.6.0","shell":"/bin/zsh","userTimeZone":"Africa/Lagos","memory":0,"machineId":"bdb7e427269f249517f77a7db88829191951f3b5a19e51e4b24e806a2a6501b8","env":"test"}}
