+++ package.json
@@ -20,22 +20,15 @@
   "author": "Justice Ogbonna",
   "license": "ISC",
   "devDependencies": {
-    "@types/geoip-lite": "^1.4.4",
     "@types/jest": "^29.5.13",
     "@types/node": "^22.7.5",
-    "esm": "^3.2.25",
-    "jest": "^29.7.0",
     "ts-jest": "^29.2.5",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.6.3"
+    "ts-node": "^10.9.2"
   },
   "dependencies": {
-    "@types/axios": "^0.14.0",
-    "axios": "^1.7.7",
+    "@google/generative-ai": "^0.21.0",
     "dotenv": "^16.4.5",
-    "geoip-lite": "^1.4.10",
-    "glob": "^11.0.0",
-    "openai": "^4.67.3",
-    "url": "^0.11.4"
+    "node-machine-id": "^1.1.12",
+    "openai": "^4.67.3"
   }
 }
